Implement ITimelineEventProvider to show git tags on timeline.
Adds the ITimlineEventProvider methods to GitConnector.  Currently
queries all Git repos for tags (could potentially limit to a subset
of repos).  Should handle varying timezones correctly (not tested).

Creates timeline entries like
  "TagName created in RepoName by User"
with the tag message on the next line.  The timeline entry should
be a link to the object pointed to by the Git tag.