Offset duplicate changeset timestamps by small amount
Several git operations can cause multiple commits to have the same
timestamp.  The most common is a rebase, which leaves the author time
alone, but sets the committer time to the rebase time.  This change
inserts microsecond differences between subsequent commits that have the
same timestamp.  It assumes the commits are encountered in the correct
topological order.  The time offset grows with each commit at the same
time, but is reset each time a new timestamp comes up.